requirements.txt
Django>=5.0.6
djangorestframework>=3.15.1
ytmusicapi>=1.3.2
Pillow>=10.3.0
requests>=2.31.0
python-decouple>=3.8  # Optional, for environment variables
